+++ web-service/src/main/java/io/acari/handler/data/AllPageHandler.java
@@ -1,80 +1,12 @@
 package io.acari.handler.data;
 
-import io.acari.util.ChainableOptional;
-import io.vertx.core.Handler;
-import io.vertx.core.eventbus.Message;
-import io.vertx.core.json.JsonArray;
-import io.vertx.core.json.JsonObject;
-import io.vertx.ext.mongo.FindOptions;
 import io.vertx.ext.mongo.MongoClient;
-import io.vertx.rx.java.ObservableFuture;
-import io.vertx.rx.java.RxHelper;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import rx.Observable;
 
-import java.util.List;
-
-public class AllPageHandler implements Handler<Message<JsonObject>> {
-    private static final Logger LOGGER = LoggerFactory.getLogger(AllPageHandler.class);
-    private static final int PAGES_PER_PAGE = 100;
-    private final MongoClient mongoClient;
-
-    public AllPageHandler(MongoClient mongoClient) {
-        this.mongoClient = mongoClient;
-    }
+public class AllPageHandler extends BaseAllPageHandler {
 
 
-    @Override
-    public void handle(Message<JsonObject> message) {
-        ChainableOptional.ofNullable(message.body().getInteger("pageNumber"))
-                .ifPresent(pageNumber -> {
-                    Observable<JsonObject> countResult = getLongObservableFuture();
-                    Observable<JsonObject> listResult = getListObservableFuture(pageNumber);
-                    listResult.zipWith(countResult,
-                            (listRes, countRes) -> new JsonObject()
-                                    .put("pages", listRes.getJsonArray("result"))
-                                    .put("metaData", new JsonObject()
-                                            .put("itemsPerPage", PAGES_PER_PAGE)
-                                            .put("currentPageNumber", pageNumber)
-                                            .put("totalItems", countRes.getLong("result"))))
-                            .subscribe(
-                                    message::reply,
-                                    throwable -> {
-                                        LOGGER.warn("Aww snap", throwable.getMessage());
-                                        logFail(message, throwable.getMessage());
-                                    });
-                })
-                .orElseDo(() -> message.fail(ErrorCodes.BAD_ACTION.ordinal(), "No page number"));
-    }
+  public AllPageHandler(MongoClient mongoClient) {
+    super(mongoClient, "pages");
+  }
 
-    private Observable<JsonObject> getListObservableFuture(Integer pageNumber) {
-        ObservableFuture<List<JsonObject>> listObservableFuture = RxHelper.observableFuture();
-        mongoClient.findWithOptions("pages",
-                new JsonObject(),
-                new FindOptions()
-                        .setLimit(PAGES_PER_PAGE)
-                        .setSkip(getSkipCount(pageNumber)), listObservableFuture.toHandler());
-        return listObservableFuture.map(ar ->
-                ar.stream()
-                        .map(json -> json.getString("name"))
-                        .collect(JsonArray::new, JsonArray::add, JsonArray::add))
-                .map(list -> new JsonObject().put("result", list));
-    }
-
-    private Observable<JsonObject> getLongObservableFuture() {
-        ObservableFuture<Long> longObservableFuture = RxHelper.observableFuture();
-        mongoClient.count("pages",
-                new JsonObject(), longObservableFuture.toHandler());
-        return longObservableFuture.map(count -> new JsonObject().put("result", count));
-    }
-
-    private void logFail(Message<JsonObject> message, String message2) {
-        LOGGER.warn("Ohh shit", message2);
-        message.fail(ErrorCodes.DB_ERROR.ordinal(), message2);
-    }
-
-    private Integer getSkipCount(Integer pageNumber) {
-        return --pageNumber * PAGES_PER_PAGE;
-    }
 }
